// frontend/src/api-generated/models/nguoiDungRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface NguoiDungRequest {
  tenDangNhap: string;
  matKhau?: string;
  hoNguoiDung?: string;
  tenDemNguoiDung?: string;
  tenNguoiDung: string;
  chucVu?: string;
  email?: string;
  soDienThoai?: string;
  danhDaiDienUrl?: string;
  idPhongBan?: number;
  danhSachIdVaiTro?: number[];
}